test: consolidate MLA KV buffer coverage
